Move DependencyGraphUtils into its own protected file.
