+++ Editor/Dependencies/Graph/DependencyUtilities.cs
@@ -77,32 +77,4 @@ namespace UnityEditor.Search
             GUI.BeginGroup(new Rect(0.0f, kEditorWindowTabHeight, Screen.width, Screen.height));
         }
     }
-
-    static class DependencyGraphUtils
-    {
-        public static Rect GetBoundingBox(IEnumerable<Node> nodes)
-        {
-            var xMin = float.MaxValue;
-            var yMin = float.MaxValue;
-            var xMax = float.MinValue;
-            var yMax = float.MinValue;
-            foreach (var node in nodes)
-            {
-                if (node == null)
-                    continue;
-
-                if (node.rect.xMin < xMin)
-                    xMin = node.rect.xMin;
-                if (node.rect.xMax > xMax)
-                    xMax = node.rect.xMax;
-                if (node.rect.yMin < yMin)
-                    yMin = node.rect.yMin;
-                if (node.rect.yMax > yMax)
-                    yMax = node.rect.yMax;
-            }
-
-            var bb = Rect.MinMaxRect(xMin, yMin, xMax, yMax);
-            return bb;
-        }
-    }
 }
